# perf: improve constructInput's efficiency